requirements.txt
wheel
pyyaml
waitress